requirements.txt
alembic==0.9.9
APScheduler==3.5.1
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-Blogging==1.0.2
Flask-Cache==0.13.1
Flask-FileUpload==0.5.0
Flask-Login==0.4.1
Flask-Migrate==2.2.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
lxml==4.2.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
newspaper3k==0.2.6
nltk==3.3
Pillow==5.1.0
python-dateutil==2.7.3
python-editor==1.0.3
python-slugify==1.2.5
pytz==2018.4
PyYAML==3.12
requests==2.19.0
requests-file==1.4.3
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.8
tldextract==2.2.0
tzlocal==1.5.1
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
